Added logout mode. This only works with tomcat 7. Tomcat 7 is not out, but one can compile it from http://svn.apache.org/repos/asf/tomcat/trunk
It requires the tomcat-coyote.jar to be available, in order for the cast
to compile: 

SLSessionManager s =(SSLSessionManager) request.getAttribute("javax.servlet.request.ssl_session_mgr");

I was able to add that jar to my maven repository with the following command line

mvn -X install:install-file -Dfile=lib/tomcat-coyote.jar -DgroupId=tomcat7 -DartifactId=tomcat-coyote -Dversion=7-trunk -Dpackaging=jar

But then as it is now the package has to be of type jar, which means it gets included in the distribution.  If that jar is in the war file, then tomcat will load it with a different class loader, and end up giving it a different name. So that file has to later be removed from the war file.

Of course there has to be a way to do this correctly in Maven.

I just wanted to proove here that logout could succeed.



git-svn-id: https://sommer.dev.java.net/svn/sommer/foafssl/trunk@501 a925da88-a912-0410-a1af-dd118eaf9e3a